#14 - 주문하기 기능에 필요한 Product 도메인 수정
- 상품 필드에 수량 추가
- id 타입 int ->long 변경

+++ src/main/java/com/example/onlinemarket/domain/product/controller/ProductController.java
@@ -59,7 +59,7 @@ public class ProductController {
     public ResponseEntity<ApiResponse<ProductDTO>> createProduct(
         @RequestBody @Valid ProductDTO productDTO) {
 
-        int productId = productService.createProduct(productDTO);
+        long productId = productService.createProduct(productDTO);
 
         URI location = ServletUriComponentsBuilder
             .fromCurrentRequest().path("/{id}")
